requirements.txt
anthropic
flask
python-dotenv
gunicorn
